[config] Refactor tests to have common base
This is to make sure different config implementation have the same
base tests to make sure they behave the same
